backend/cloudfunctions/requirements.txt
functions-framework==3.*
gspread==6.1.*
google-cloud-storage==2.18.*
